🔊 Overhaul console handling using a temporary file
Printing console messages is actually tricky. When fzf controls the screen,
we would need to switch terminal screens print to main screen, but that causes
an annoying flicker. And that does not work for reload commands, whose stdout
and stderr are redirected by fzf.

The practical solution is to write all console output to a temporary file and
print it out on exit. This is coordinated by the TUICK_LOG_FILE environment
variable: if a process does not see it set, it creates temporary file, its path
to TUICK_LOG_FILE, redirect the console, and print the full output on exit.
When a subprocess sees TUICK_LOG_FILE set, it just redirects its console to the
file.